requirements.txt
python-telegram-bot==21.4
ccxt==4.1.13
pydantic>=2.8
pydantic-settings>=2.4
python-dotenv>=1.0
httpx>=0.27
tenacity>=8.5
aiolimiter>=1.1
portalocker>=2.10
typer>=0.12
uvloop>=0.19; platform_system=="Linux"
prometheus-client>=0.20
orjson>=3.10